rename setupModules folder to SetupModulesAndDeck
